changed BuildScript to be IPreprocessBuildWithReport
