+++ Explorer/Assets/Editor/BuildScript.cs
@@ -1,260 +1,81 @@
-﻿using DCL.PerformanceAndDiagnostics.Analytics;
+﻿
+using DCL.PerformanceAndDiagnostics.Analytics;
 using JetBrains.Annotations;
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEditor;
+using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
 using UnityEngine;
 
 namespace Editor
 {
-    public static class BuildScript
+    public class BuildScript : IPreprocessBuildWithReport
     {
-        private static readonly string Eol = Environment.NewLine;
+        [PublicAPI]
+        public static Dictionary<string, object> Parameters { get; private set; }
 
-        private static readonly string[] Secrets =
-            {"androidKeystorePass", "androidKeyaliasName", "androidKeyaliasPass"};
+        public int callbackOrder { get; }
 
-        [UsedImplicitly]
-        public static void Build()
+        public void OnPreprocessBuild(BuildReport report)
         {
+            Debug.Log("~~ OnPreprocessBuild ~~");
+
+            // Get all environment variables
+            var environmentVariables = Environment.GetEnvironmentVariables();
+            Parameters = environmentVariables.Cast<DictionaryEntry>().ToDictionary(x => x.Key.ToString(), x => x.Value);
+
+            // E.g. access like:
+            // Debug.Log(Parameters["TEST_VALUE"] as string);
+
             //Unity suggestion: 1793168
             //This should ensure that the rosyln compiler has been ran and everything is generated as needed.
             EditorApplication.ExecuteMenuItem("File/Save Project");
 
-            // Gather values from args
-            Dictionary<string, string> options = GetValidatedOptions();
-
             // Set version for this build
-            PlayerSettings.bundleVersion = options["buildVersion"];
-            PlayerSettings.macOS.buildNumber = options["buildVersion"];
-            PlayerSettings.Android.bundleVersionCode = int.Parse(options["androidVersionCode"]);
+            var buildVersion = Parameters["BUILD_VERSION"] as string;
+            PlayerSettings.bundleVersion = buildVersion;
+            PlayerSettings.macOS.buildNumber = buildVersion;
+            Debug.Log($"Build version set to: {buildVersion}");
 
-            // Apply build target
-            var buildTarget = (BuildTarget) Enum.Parse(typeof(BuildTarget), options["buildTarget"]);
-            switch (buildTarget)
-            {
-                case BuildTarget.Android:
-                {
-                    EditorUserBuildSettings.buildAppBundle = options["customBuildPath"].EndsWith(".aab");
-                    if (options.TryGetValue("androidKeystoreName", out string keystoreName) &&
-                        !string.IsNullOrEmpty(keystoreName))
-                    {
-                      PlayerSettings.Android.useCustomKeystore = true;
-                      PlayerSettings.Android.keystoreName = keystoreName;
-                    }
-                    if (options.TryGetValue("androidKeystorePass", out string keystorePass) &&
-                        !string.IsNullOrEmpty(keystorePass))
-                        PlayerSettings.Android.keystorePass = keystorePass;
-                    if (options.TryGetValue("androidKeyaliasName", out string keyaliasName) &&
-                        !string.IsNullOrEmpty(keyaliasName))
-                        PlayerSettings.Android.keyaliasName = keyaliasName;
-                    if (options.TryGetValue("androidKeyaliasPass", out string keyaliasPass) &&
-                        !string.IsNullOrEmpty(keyaliasPass))
-                        PlayerSettings.Android.keyaliasPass = keyaliasPass;
-                    if (options.TryGetValue("androidTargetSdkVersion", out string androidTargetSdkVersion) &&
-                        !string.IsNullOrEmpty(androidTargetSdkVersion))
-                    {
-                        var targetSdkVersion = AndroidSdkVersions.AndroidApiLevelAuto;
-                        try
-                        {
-                            targetSdkVersion =
-                                (AndroidSdkVersions) Enum.Parse(typeof(AndroidSdkVersions), androidTargetSdkVersion);
-                        }
-                        catch
-                        {
-                            Debug.Log("Failed to parse androidTargetSdkVersion! Fallback to AndroidApiLevelAuto");
-                        }
-
-                        PlayerSettings.Android.targetSdkVersion = targetSdkVersion;
-                    }
-
-                    break;
-                }
-                case BuildTarget.StandaloneOSX:
-                    PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, ScriptingImplementation.Mono2x);
-                    break;
-            }
-
-            // Determine subtarget
-            int buildSubtarget = 0;
-#if UNITY_2021_2_OR_NEWER
-            if (!options.TryGetValue("standaloneBuildSubtarget", out var subtargetValue) || !Enum.TryParse(subtargetValue, out StandaloneBuildSubtarget buildSubtargetValue)) {
-                buildSubtargetValue = default;
-            }
-            buildSubtarget = (int) buildSubtargetValue;
-#endif
-
-            // Custom build
-            Build(buildTarget, buildSubtarget, options);
+            var segmentKey = Parameters["SEGMENT_WRITE_KEY"] as string;
+            Debug.Log($"[SEGMENT]: write key found");
+            WriteSegmentKeyToAnalyticsConfig(segmentKey);
         }
 
-        private static Dictionary<string, string> GetValidatedOptions()
+        private static void WriteSegmentKeyToAnalyticsConfig(string segmentWriteKey)
         {
-            ParseCommandLineArguments(out Dictionary<string, string> validatedOptions);
+            string[] guids = AssetDatabase.FindAssets($"t:{nameof(AnalyticsConfiguration)}");
 
-            if (!validatedOptions.TryGetValue("projectPath", out string _))
+            switch (guids.Length)
             {
-                Console.WriteLine("Missing argument -projectPath");
-                EditorApplication.Exit(110);
-            }
-
-            if (!validatedOptions.TryGetValue("buildTarget", out string buildTarget))
-            {
-                Console.WriteLine("Missing argument -buildTarget");
-                EditorApplication.Exit(120);
-            }
-
-            if (!Enum.IsDefined(typeof(BuildTarget), buildTarget ?? string.Empty))
-            {
-                Console.WriteLine($"{buildTarget} is not a defined {nameof(BuildTarget)}");
-                EditorApplication.Exit(121);
-            }
-
-            if (!validatedOptions.TryGetValue("customBuildPath", out string _))
-            {
-                Console.WriteLine("Missing argument -customBuildPath");
-                EditorApplication.Exit(130);
-            }
-
-            const string defaultCustomBuildName = "TestBuild";
-            if (!validatedOptions.TryGetValue("customBuildName", out string customBuildName))
-            {
-                Console.WriteLine($"Missing argument -customBuildName, defaulting to {defaultCustomBuildName}.");
-                validatedOptions.Add("customBuildName", defaultCustomBuildName);
-            }
-            else if (customBuildName == "")
-            {
-                Console.WriteLine($"Invalid argument -customBuildName, defaulting to {defaultCustomBuildName}.");
-                validatedOptions.Add("customBuildName", defaultCustomBuildName);
-            }
-
-            return validatedOptions;
-        }
-
-        private static void ParseCommandLineArguments(out Dictionary<string, string> providedArguments)
-        {
-            providedArguments = new Dictionary<string, string>();
-            string[] args = Environment.GetCommandLineArgs();
-
-            Console.WriteLine(
-                $"{Eol}" +
-                $"###########################{Eol}" +
-                $"#    Parsing settings     #{Eol}" +
-                $"###########################{Eol}" +
-                $"{Eol}"
-            );
-
-            // Extract flags with optional values
-            for (int current = 0, next = 1; current < args.Length; current++, next++)
-            {
-                // Parse flag
-                bool isFlag = args[current].StartsWith("-");
-                if (!isFlag) continue;
-                string flag = args[current].TrimStart('-');
-
-                // Parse optional value
-                bool flagHasValue = next < args.Length && !args[next].StartsWith("-");
-                string value = flagHasValue ? args[next].TrimStart('-') : "";
-                bool secret = Secrets.Contains(flag);
-                string displayValue = secret ? "*HIDDEN*" : "\"" + value + "\"";
-
-                // Assign
-                Console.WriteLine($"Found flag \"{flag}\" with value {displayValue}.");
-                providedArguments.Add(flag, value);
-            }
-        }
-
-        private static void Build(BuildTarget buildTarget, int buildSubtarget, Dictionary<string, string> options)
-        {
-            string[] scenes = EditorBuildSettings.scenes.Where(scene => scene.enabled).Select(s => s.path).ToArray();
-            var buildPlayerOptions = new BuildPlayerOptions
-            {
-                scenes = scenes,
-                target = buildTarget,
-                locationPathName = options["customBuildPath"],
-#if UNITY_2021_2_OR_NEWER
-                subtarget = buildSubtarget
-#endif
-            };
-
-            buildPlayerOptions.options |=  BuildOptions.DetailedBuildReport;
-
-            bool isProfilingBuild = options.TryGetValue("profile", out string _);
-            bool isDeepProfilingBuild = options.TryGetValue("deepProfile", out string _);
-
-            Console.WriteLine($"Profiling Enabled: {isProfilingBuild}, Deep Profiling Enabled: {isDeepProfilingBuild}");
-
-            string enableProfilingDefine = "ENABLE_PROFILING";
-            if (isProfilingBuild || isDeepProfilingBuild)
-            {
-                buildPlayerOptions.options |= BuildOptions.ConnectWithProfiler;
-                buildPlayerOptions.options |= BuildOptions.Development;
-
-                if(buildPlayerOptions.extraScriptingDefines == null)
-                    buildPlayerOptions.extraScriptingDefines = new[] {enableProfilingDefine};
-                else
-                {
-                    var currentDefines = buildPlayerOptions.extraScriptingDefines.ToList();
-                    currentDefines.Add(enableProfilingDefine);
-                    buildPlayerOptions.extraScriptingDefines = currentDefines.ToArray();
-                }
-            }
-
-            if (isDeepProfilingBuild)
-            {
-                buildPlayerOptions.options |= BuildOptions.EnableDeepProfilingSupport;
-            }
-
-            BuildSummary buildSummary = BuildPipeline.BuildPlayer(buildPlayerOptions).summary;
-            ReportSummary(buildSummary);
-            ExitWithResult(buildSummary.result);
-        }
-
-        private static void ReportSummary(BuildSummary summary)
-        {
-            Debug.Log("Logging fake Build results so that the build via game-ci/unity-builder does not fail...");
-            Debug.Log($"###########################{Environment.NewLine}#      Build results      #{Environment.NewLine}###########################{Environment.NewLine}" +
-                      $"{Environment.NewLine}Duration: 00:00:00.0000000{Environment.NewLine}Warnings: 0{Environment.NewLine}Errors: 0{Environment.NewLine}Size: 0 bytes{Environment.NewLine}{Environment.NewLine}{Environment.NewLine}Build succeeded!");
-            /*
-            Console.WriteLine(
-                $"{Eol}" +
-                $"###########################{Eol}" +
-                $"#      Build results      #{Eol}" +
-                $"###########################{Eol}" +
-                $"{Eol}" +
-                $"Duration: {summary.totalTime.ToString()}{Eol}" +
-                $"Warnings: {summary.totalWarnings.ToString()}{Eol}" +
-                $"Errors: {summary.totalErrors.ToString()}{Eol}" +
-                $"Size: {summary.totalSize.ToString()} bytes{Eol}" +
-                $"{Eol}"
-            );
-            */
-        }
-
-        private static void ExitWithResult(BuildResult result)
-        {
-            switch (result)
-            {
-                case BuildResult.Succeeded:
-                    //Console.WriteLine("Build succeeded!");
-                    EditorApplication.Exit(0);
-                    break;
-                case BuildResult.Failed:
-                    //Console.WriteLine("Build failed!");
-                    EditorApplication.Exit(101);
-                    break;
-                case BuildResult.Cancelled:
-                    //Console.WriteLine("Build cancelled!");
-                    EditorApplication.Exit(102);
-                    break;
-                case BuildResult.Unknown:
-                default:
-                    //Console.WriteLine("Build result is unknown!");
-                    EditorApplication.Exit(103);
+                case 0:
+                    Debug.LogError($"{nameof(AnalyticsConfiguration)} asset not found!");
+                    return;
+                case > 1:
+                    Debug.LogWarning($"Multiple {nameof(AnalyticsConfiguration)} assets found. Using the first one.");
                     break;
             }
+
+            AssetDatabase.Refresh();
+            string assetPath = AssetDatabase.GUIDToAssetPath(guids[0]);
+            AnalyticsConfiguration config = AssetDatabase.LoadAssetAtPath<AnalyticsConfiguration>(assetPath);
+
+            if (config != null)
+            {
+                Debug.Log($"[SEGMENT]: write key lenght {segmentWriteKey.Length}");
+
+                config.SetWriteKey(segmentWriteKey);
+                EditorUtility.SetDirty(config);
+                AssetDatabase.SaveAssets();
+                AssetDatabase.Refresh();
+
+                Debug.Log("[SEGMENT]: write key saved");
+            }
+            else // TODO (Vit): create default and add to Addressables (config = ScriptableObject.CreateInstance<AnalyticsConfiguration>());
+                Debug.LogWarning($"{nameof(AnalyticsConfiguration)} asset not found , when trying to load it from AssetDatabase. Creating SO config file via {nameof(ScriptableObject.CreateInstance)}");
         }
     }
 }
