+++ Explorer/Assets/Editor/BuildScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 40cac8947048efb4996d7c24886feda7
+guid: 5541507c271ee43469e39ed649833593
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
