--- Explorer/Assets/Editor/CloudBuild.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5541507c271ee43469e39ed649833593
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
